+++ src/Symfony/Component/Form/Extension/Core/Type/TextareaType.php
@@ -20,7 +20,7 @@ class TextareaType extends AbstractType
     /**
      * {@inheritdoc}
      */
-    public function buildView(FormView $view, FormInterface $form)
+    public function buildView(FormView $view, FormInterface $form, array $options)
     {
         $view->set('pattern', null);
     }